server/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==4.1.2; python_version >= '3.7'
beautifulsoup4==4.12.2; python_full_version >= '3.6.0'
blinker==1.7.0; python_version >= '3.8'
certifi==2023.11.17; python_full_version >= '3.6.0'
charset-normalizer==3.3.2; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
extruct==0.16.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask==3.0.0
gunicorn==21.2.0
html-text==0.5.2
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.6; python_version >= '3.5'
isodate==0.6.1
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jstyleson==0.0.2
lxml==4.9.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==2.1.3; python_version >= '3.7'
mf2py==2.0.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
psycopg2-binary==2.9.9
pymysql==1.1.0
pyparsing==3.1.1; python_full_version >= '3.6.8'
pyrdfa3==3.5.3
python-dotenv==1.0.0
rdflib==7.0.0; python_version >= '3.7'
recipe-scrapers==14.53.0
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'
w3lib==2.1.2; python_version >= '3.7'
webencodings==0.5.1
werkzeug==3.0.1; python_version >= '3.8'
